Re #169 separated hpc_config from hor_config
git-svn-id: https://svn.isis.rl.ac.uk/Horace/branches/Re169_generic_mpi@1584 1cb955da-4463-431d-88dd-2af2f5550de3

+++ configuration/@hor_config/hor_config.m
@@ -31,20 +31,6 @@ classdef hor_config<config_base
     %   delete_tmp        -  automatically delete temporary files after generating sqw files
     %   working_directory - the folder to write tmp files.
     %
-    %-----
-    % Methods related to parallel file processing/combining on a cluster:
-    % use_mex_for_combine     - use mex code to combine various sqw/tmp files
-    %                           together
-    % mex_combine_thread_mode - various thread modes deployed when
-    %                           combining sqw files using mex code.
-    % mex_combine_buffer_size - size of buffer used by mex code while
-    %                           combining files per each contributing file.
-    % accum_in_separate_process - if true, launch separate Matlab session(s) to generate tmp files
-    % accumulating_process_num  - number of Matlab sessions to launch to calculate tmp files
-    %
-    %
-    %
-    % Type >> hor_config  to see the list of current configuration option values.
     %
     % $Revision$ ($Date$)
     %
@@ -89,30 +75,10 @@ classdef hor_config<config_base
         % Assign empty value to restore it to default (system tmp
         % directory)
         working_directory
-        % use multi-threaded mex code to combine various sqw/tmp files together
-        use_mex_for_combine
-        % various thread modes deployed when combining sqw files using mex code:
-        % namely:
-        % 0  - one thread read all tmp files and another one writes combined
-        %      information into the target file
-        % 1  - one thread writes combined sqw file and two threads are
-        %      launched for each contributing file to read necessary information.
-        %      mode 1 is combinations of modes 2 and 3.
-        % Two debug modes exist to separate reading of this information:
-        % 2  - a thread per contributing file is launched to read bin information
-        %      when common thread for all contributing files is used to
-        %      read pixel information
-        % 3  - a thread is launched per contributing file to read pixel
-        %      information while common thread is used to read bin
-        %      information
-        mex_combine_thread_mode
-        % size of buffer used by mex code while combining files per each
-        % file.
-        mex_combine_buffer_size
-        % if true, launch separate Matlab session(s) to generate tmp files
-        accum_in_separate_process
-        % number of sessions to launch to calculate additional files
-        accumulating_process_num
+        % the property, related to high performance computing settings.
+        % Here it provided for information only while changes to this
+        % propery should be made through hpc_config directly.
+        hpc_config_info
     end
     properties(Dependent,Hidden)
         % old implementation of log_level property
@@ -135,14 +101,6 @@ classdef hor_config<config_base
         force_mex_if_use_mex_ = false;
         delete_tmp_ = true;
         working_directory_ ='';
-        
-        use_mex_for_combine_ = false;
-        %
-        mex_combine_thread_mode_   = 0;
-        mex_combine_buffer_size_ = 1024*64;
-        
-        accum_in_separate_process_ = false;
-        accumulating_process_num_ = 2;
     end
     
     properties(Constant,Access=private)
@@ -151,9 +109,7 @@ classdef hor_config<config_base
         saved_properties_list_={'mem_chunk_size','threads','ignore_nan',...
             'ignore_inf', 'log_level','use_mex',...
             'force_mex_if_use_mex','delete_tmp',...
-            'working_directory',...
-            'mex_combine_thread_mode','mex_combine_buffer_size','use_mex_for_combine',...
-            'accum_in_separate_process','accumulating_process_num'}
+            'working_directory'}
     end
     
     methods
@@ -162,15 +118,6 @@ classdef hor_config<config_base
             this=this@config_base(mfilename('class'));
             
             this.threads_ = find_nproc_to_use(this);
-            % set os-specific defaults
-            if ispc
-                this.mex_combine_thread_mode_   = 0;
-            elseif isunix
-                if ~ismac
-                    this.mex_combine_thread_mode_   = 0;
-                    this.mex_combine_buffer_size_ = 64*1024;
-                end
-            end
         end
         
         %-----------------------------------------------------------------
@@ -202,21 +149,6 @@ classdef hor_config<config_base
         end
         function delete = get.delete_tmp(this)
             delete = get_or_restore_field(this,'delete_tmp');
-        end
-        function use = get.use_mex_for_combine(this)
-            use = get_or_restore_field(this,'use_mex_for_combine');
-        end
-        function size= get.mex_combine_buffer_size(this)
-            size = get_or_restore_field(this,'mex_combine_buffer_size');
-        end
-        function type= get.mex_combine_thread_mode(this)
-            type = get_or_restore_field(this,'mex_combine_thread_mode');
-        end
-        function accum = get.accum_in_separate_process(this)
-            accum = get_or_restore_field(this,'accum_in_separate_process');
-        end
-        function accum = get.accumulating_process_num(this)
-            accum = get_or_restore_field(this,'accumulating_process_num');
         end
         function work_dir = get.working_directory(this)
             work_dir = get_or_restore_field(this,'working_directory');
@@ -332,73 +264,6 @@ classdef hor_config<config_base
             end
             config_store.instance().store_config(this,'delete_tmp',del);
         end
-        function this = set.use_mex_for_combine(this,val)
-            if val>0
-                try
-                    ver = combine_sqw();
-                    config_store.instance().store_config(this,'use_mex_for_combine',true);
-                catch ME
-                    warning('HOR_CONFIG:use_mex_for_combine',...
-                        [' combine_sqw.mex procedure is not availible.\n',...
-                        ' Reason: %s\n.',...
-                        ' Will not use mex for combininng'],ME.message);
-                    config_store.instance().store_config(this,'use_mex_for_combine',false);
-                end
-            else
-                config_store.instance().store_config(this,'use_mex_for_combine',false);
-            end
-        end
-        function this= set.mex_combine_buffer_size(this,val)
-            if val<64
-                error('HOR_CONFIG:mex_combine_buffer_size',...
-                    ' mex_combine_buffer_size should be bigger then 64, and better >1024');
-            end
-            if val==0
-                this.use_mex_for_combine = false;
-                return;
-            end
-            config_store.instance().store_config(this,'mex_combine_buffer_size',val);
-        end
-        function this= set.mex_combine_thread_mode(this,val)
-            if  val>3 || val < 0
-                error('HOR_CONFIG:mex_combine_thread_mode',...
-                    [' mex_combine_multithreaded should be a number in the range fromn 0 to 3\n ',...
-                    '  meaning:\n', ...
-                    ' 0 -- minor multitheading ',...
-                    ' 1 -- full multitrheading',...
-                    ' and two debug options:\n', ...
-                    ' 2 -- only bin numbers are read by separate thread',...
-                    ' 3 -- only pixels are read by separate thread']);
-            end
-            config_store.instance().store_config(this,'mex_combine_thread_mode',val);
-        end
-        function this = set.accum_in_separate_process(this,val)
-            if val>0
-                accum = true;
-            else
-                accum = false;
-            end
-            if accum
-                [ok,mess] = check_worker_configured();
-                if ~ok
-                    warning('HOR_CONFIG:set_accum_in_separate_process',...
-                        ' Can not start accumulating in separate process as: %s',...
-                        mess);
-                    accum = false;
-                end
-            end
-            config_store.instance().store_config(this,'accum_in_separate_process',accum);
-            
-        end
-        function this = set.accumulating_process_num(this,val)
-            if val<1
-                error('HOR_CONFIG:accumulating_process_num',...
-                    'Number of accumulating processes should be more then 1');
-            else
-                nproc = val;
-            end
-            config_store.instance().store_config(this,'accumulating_process_num',nproc);
-        end
         function this = set.working_directory(this,val)
             % Check and set working directory
             if ~is_string(val)
@@ -406,14 +271,18 @@ classdef hor_config<config_base
                     'working directory value should be a ')
             end
             if ~isempty(val)
-                test_dir = fullfile(val,'horace_test_write_directory');
-                clob = onCleanup(@()rmdir(test_dir,'s'));
-                ok = mkdir(test_dir);
-                if ~ok
-                    warning('HOR_CONFIG:invalid_argument',...
-                        'working directory %s does not have write permissions. Changing it to %s directory',...
-                        val,tempdir);
+                if strcmp(val,tempdir) % avoid storing tmp dir as working directory as this is default
                     val = '';
+                else
+                    test_dir = fullfile(val,'horace_test_write_directory');
+                    clob = onCleanup(@()rmdir(test_dir,'s'));
+                    ok = mkdir(test_dir);
+                    if ~ok
+                        warning('HOR_CONFIG:invalid_argument',...
+                            'working directory %s does not have write permissions. Changing it to %s directory',...
+                            val,tempdir);
+                        val = '';
+                    end
                 end
             end
             config_store.instance().store_config(this,'working_directory',val);
